Checkpoint
Thought this lab was gonna be hard, no it seems like the UML will be the hardest part. Just due to it being new, so, with that saids still sprinting to get it done but not as dire as i thought.

So, called the toString wit ha simple for loop and then added in the menu to print at least. Doesnt accept input actually yet